services/trade_producer/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Devopdevs <owoeyebabatunde01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
quixstreams = "^2.11.1"
make = "^0.1.6.post2"
loguru = "^0.7.2"
websocket-client = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
